added letter and game class

+++ src/Box.java
@@ -29,7 +29,7 @@ public class Box
     }
 
 
-    public boolean anySideLongerThan(int side)
+    public boolean anySideLongerThan(double side)
     {
         if (length > side || width > side || height > side)
         {
